Add Elixir name service
This is another vanilla plug + otel collector service, but now it calls
out over HTTP to the year service. I'd like to use the telepoison
library to handle this automatically, but its dependency on otel is
declared at ~> 0.5, which interferes with opentelemetry_plug. So I'll
just manually inject the HTTPoison headers when I figure out how the
propagation is configured.