add config option to allow redis reconnect attempts
This parses in the REDIS_URL string and then adds to the parsed options
a setting to allow connection failures. Connection attempts still throw
an exception, but the process stays up. The result will be if Redis isn't
available, the .NET Message service will return an HTTP/500 (and
currently blats out the exception in the response body).