<2.x of otel-java just needs env var lol
removing log4j2.xml bc its not used at all.
log4j2-spring.xml would be used and it
would require some changes to log pkgs
so for now we'll just keep defaults.

if using the sdk instead of the java agent,
this work would be required, as would
the other packages like the appender.